text-comparison-system/requirements.txt
Flask>=2.0
python-dotenv>=0.19
nltk>=3.6
scikit-learn>=1.0
PyPDF2>=3.0.0
python-docx>=1.1.0